+++ marketplace/services/product/product_facebook_manage.py
@@ -3,7 +3,11 @@ from typing import List
 from django.contrib.auth import get_user_model
 from django.db import transaction
 
-from marketplace.wpp_products.models import Product, UploadProduct, Catalog, ProductFeed
+from marketplace.wpp_products.models import (
+    UploadProduct,
+    Catalog,
+    ProductFeed,
+)
 from marketplace.services.vtex.utils.data_processor import (
     DataProcessor,
     FacebookProductDTO,
@@ -14,93 +18,6 @@ User = get_user_model()
 
 
 class ProductFacebookManager:
-    def create_or_update_products_on_database(
-        self, products: List[FacebookProductDTO], catalog, product_feed
-    ):
-        print("Starting the data insertion process in integrations databases")
-
-        # Initialize dictionaries to map existing products
-        existing_products_map = {}
-
-        # Search all relevant products at once based on the specified catalog
-        existing_products_qs = Product.objects.filter(
-            facebook_product_id__in=[dto.id for dto in products], catalog=catalog
-        )
-
-        # Update the mapping using the combination of facebook_product_id and catalog_id
-        for product in existing_products_qs:
-            existing_products_map[
-                (product.facebook_product_id, product.catalog_id)
-            ] = product
-
-        products_to_update = []
-        products_to_create = []
-
-        for dto in products:
-            key = (dto.id, catalog.id)  # Combination of ID and catalog_id as key
-            existing_product = existing_products_map.get(key)
-
-            if existing_product:
-                # Update existing product
-                existing_product.title = dto.title
-                existing_product.description = dto.description
-                existing_product.availability = dto.availability
-                existing_product.condition = dto.condition
-                existing_product.price = dto.price
-                existing_product.link = dto.link
-                existing_product.image_link = dto.image_link
-                existing_product.brand = dto.brand
-                existing_product.sale_price = dto.sale_price
-                products_to_update.append(existing_product)
-            else:
-                # Prepare a new product for creation
-                new_product = Product(
-                    facebook_product_id=dto.id,
-                    title=dto.title,
-                    description=dto.description,
-                    availability=dto.availability,
-                    condition=dto.condition,
-                    price=dto.price,
-                    link=dto.link,
-                    image_link=dto.image_link,
-                    brand=dto.brand,
-                    sale_price=dto.sale_price,
-                    catalog=catalog,
-                    created_by=User.objects.get_admin_user(),
-                    feed=product_feed,
-                )
-                products_to_create.append(new_product)
-
-        # Bulk processing to update and create
-        with transaction.atomic():
-            if products_to_update:
-                Product.objects.bulk_update(
-                    products_to_update,
-                    [
-                        "title",
-                        "description",
-                        "availability",
-                        "condition",
-                        "price",
-                        "link",
-                        "image_link",
-                        "brand",
-                        "sale_price",
-                    ],
-                    batch_size=10000,
-                )
-                print(
-                    f"{len(products_to_update)} products have been updated in the integration databases"
-                )
-
-            if products_to_create:
-                Product.objects.bulk_create(products_to_create, batch_size=10000)
-                print(
-                    f"{len(products_to_create)} products have been created in the integration databases"
-                )
-
-        return True
-
     def save_csv_product_data(
         self,
         products_dto: List[FacebookProductDTO],
@@ -129,6 +46,10 @@ class ProductFacebookManager:
             except Exception as e:
                 print(f"Failed to save or update product: {str(e)}")
                 all_success = False
+
+        # Call duplicate cleanup after processing
+        UploadProduct.remove_duplicates(catalog)
+
         print(
             f"All {len(products_dto)} products were saved successfully in the database:"
             f"Catalog ID {catalog.facebook_catalog_id}, "
@@ -179,6 +100,9 @@ class ProductFacebookManager:
             UploadProduct.objects.bulk_create(products_to_create, batch_size=batch_size)
             print(f"Bulk created {len(products_to_create)} remaining products.")
 
+        # Call duplicate cleanup after processing
+        UploadProduct.remove_duplicates(catalog)
+
         print("Initial data insertion process completed successfully.")
         return True
 
@@ -194,20 +118,20 @@ class ProductFacebookManager:
             f"Starting insertion process for {len(products_dto)} products. Catalog: {catalog.name}"
         )
 
-        # Prepare lists for bulk_create and bulk_update
-        new_products = []
-        update_products = []
-
-        # Get all existing products in the database
+        # Fetch products matching facebook_product_id from products_dto, regardless of status
         existing_products = UploadProduct.objects.filter(
             facebook_product_id__in=[product.id for product in products_dto],
             catalog=catalog,
             feed=product_feed,
         )
+
         # Create a dictionary of existing products for easy access
         existing_products_dict = {
             product.facebook_product_id: product for product in existing_products
         }
+
+        new_products = []
+        update_products = []
 
         for product in products_dto:
             facebook_product_id = product.id
@@ -230,6 +154,7 @@ class ProductFacebookManager:
                         status="pending",
                     )
                 )
+
         all_success = True
         try:
             with transaction.atomic():
@@ -243,7 +168,7 @@ class ProductFacebookManager:
                     )
 
             print(
-                f"All {len(products_dto)} products were saved successfully in the database:"
+                f"All {len(products_dto)} products were saved successfully in the database."
             )
             print(
                 f"New products: {len(new_products)}, Updateds : {len(update_products)}"
@@ -255,5 +180,8 @@ class ProductFacebookManager:
         except Exception as e:
             print(f"Failed to save or update products: {str(e)}")
             all_success = False
+
+        # Call duplicate cleanup after processing
+        UploadProduct.remove_duplicates(catalog)
 
         return all_success