+++ marketplace/services/vtex/business/rules/exclude_alcoholic_drinks.py
@@ -14,7 +14,14 @@ class ExcludeAlcoholicDrinks(Rule):
         ALCOHOLIC_DRINKS_CATEGORIES (set): A set of category names that identify alcoholic drinks.
     """
 
-    ALCOHOLIC_DRINKS_CATEGORIES = {"bebida alcoólica", "bebidas alcoólicas", "vinos y licores"}
+    ALCOHOLIC_DRINKS_CATEGORIES = {
+        "bebida alcoólica",
+        "bebidas alcoólicas",
+        "bebidas alcohólicas",
+        "vinos y licores",
+        "licores",
+        "vinos",
+    }
 
     def apply(self, product: FacebookProductDTO, **kwargs) -> bool:
         """