+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+
+v3.9.9
+----------
+* Remove UploadProduct constraint and create classmethod to UploadProduct model
+
+v3.9.8
+----------
+* feat: add view to return vtex integration details by project UUID
+* Fix Unit Tests And Refactor Request class
+* Groups log erros to facebook sync tasks
+* Adds new exclusion rules by categories to colombia
+
+v3.9.7
+----------
+* Vtex ads switch
+
 v3.9.6
 ----------
 * Add ProjectManagePermission to TemplateMessageViewSet